♻️ refactor(backend): migrar namespace gob a com.transparencia
Se detectaron clases y repositorios bajo pe.gob.saip.backend coexistiendo con com.transparencia.api, lo que introducía duplicidad de paquetes y riesgo de inconsistencias en el backend.

- Migrar entidades, repositorios y servicios desde pe.gob.saip.backend hacia com.transparencia.api
- Eliminar la ruta pe/gob del backend y consolidar imports en el namespace oficial
- Ajustar mapeos JPA heredados para mantener compatibilidad del modelo actual
- Agregar SolicitudServiceTest como cobertura unitaria de HU-02 BE-05
- Verificacion: mvnw.cmd test OK (10 tests), mvnw.cmd -DskipTests package OK
