+++ transparencia-backend/src/main/java/com/transparencia/api/repository/CiudadanoRepository.java
@@ -4,6 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.transparencia.api.model.entity.Ciudadano;
 
+import java.util.Optional;
+
 @Repository
 public interface CiudadanoRepository extends JpaRepository<Ciudadano, Long> {
+    Optional<Ciudadano> findByDni(String dni);
 }
