+++ transparencia-backend/src/main/java/com/transparencia/api/service/MiembroTTAIPService.java
@@ -1,9 +1,9 @@
-package pe.gob.saip.backend.service;
+package com.transparencia.api.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pe.gob.saip.backend.model.entity.MiembroTTAIP;
-import pe.gob.saip.backend.repository.MiembroTTAIPRepository;
+import com.transparencia.api.model.entity.MiembroTTAIP;
+import com.transparencia.api.repository.MiembroTTAIPRepository;
 
 import java.util.List;
 import java.util.Optional;
