--- transparencia-backend/src/main/java/pe/gob/saip/backend/model/entity/MiembroTTAIP.java
@@ -1,46 +1,0 @@
-package pe.gob.saip.backend.model.entity;
-
-import com.transparencia.api.model.entity.Resolucion;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.PrimaryKeyJoinColumn;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-@Entity
-@Table(name = "miembros_ttaip")
-@PrimaryKeyJoinColumn(name = "id_usuario")
-public class MiembroTTAIP extends Usuario {
-
-    @NotBlank(message = "El nombre completo no puede estar vacio")
-    @Column(name = "nombre_completo", nullable = false, length = 200)
-    private String nombreCompleto;
-
-    @Column(length = 100)
-    private String cargo;
-
-    @Column(length = 30)
-    private String sala;
-
-    @Column(length = 500)
-    private String especialidad;
-
-    @OneToMany(mappedBy = "miembroTTAIP", cascade = {CascadeType.PERSIST, CascadeType.MERGE})
-    private List<Resolucion> resoluciones;
-
-    public Long getId() {
-        return this.getIdUsuario();
-    }
-}
