--- transparencia-backend/src/main/java/pe/gob/saip/backend/repository/CiudadanoRepository.java
@@ -1,12 +1,0 @@
-package pe.gob.saip.backend.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import pe.gob.saip.backend.model.entity.Ciudadano;
-
-import java.util.Optional;
-
-@Repository
-public interface CiudadanoRepository extends JpaRepository<Ciudadano, Long> {
-    Optional<Ciudadano> findByDni(String dni);
-}
